improve lighthouse results, lazy load map
