+++ src/routes/landing/contact-info/Maps.jsx
@@ -1,0 +1,34 @@
+import React from "react";
+import L from "leaflet";
+import icon from "leaflet/dist/images/marker-icon.png";
+import iconShadow from "leaflet/dist/images/marker-shadow.png";
+import { MapContainer, TileLayer } from "react-leaflet";
+import { Marker } from "react-leaflet";
+
+let DefaultIcon = L.icon({
+  iconUrl: icon,
+  shadowUrl: iconShadow,
+});
+L.Marker.prototype.options.icon = DefaultIcon;
+const position = [37.0902, -93.7129];
+
+const Maps = () => {
+  return (
+    <div className="map" aria-label="Company Location">
+      <MapContainer
+        id="map"
+        center={position}
+        zoom={9}
+        scrollWheelZoom={false}
+        loading="lazy">
+        <TileLayer
+          attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
+          url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+        />
+        <Marker position={position}></Marker>
+      </MapContainer>
+    </div>
+  );
+};
+
+export default Maps;
