+++ src/routes/landing/our-service/our-service.css
@@ -1,5 +1,3 @@
-/* Section 3 */
-
 .homepage__services {
   text-align: center;
   background-color: var(--light-black);
