+++ vite.config.js
@@ -1,14 +1,14 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
 
 export default defineConfig({
-  base: '/',
+  base: "/",
   plugins: [react()],
   server: {
-    historyApiFallback: true
+    historyApiFallback: true,
   },
   css: {
-    minify: 'esbuild',
+    minify: "esbuild",
   },
   build: {
     cssCodeSplit: true,
